Update the post auto functions.
