+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FC_POST_AUTO_60_SPEED.java
@@ -30,16 +30,17 @@
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.hardware.rev.RevBlinkinLedDriver;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.hardware.sparkfun.SparkFunOTOS;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.hardware.Servo;
+import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.hardware.IMU;
+import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
@@ -102,6 +103,7 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
     boolean yPressed;
     boolean aPressed;
     boolean depositMode;
+    boolean hangingMode = false;
     boolean bPressed;
     boolean bucketMode;
     boolean guidePressed;
@@ -114,17 +116,19 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
 
     // Some constant values
     final double FRONT_CLAW_OPENED = 0.1;
-    final double FRONT_CLAW_CLOSED = 0.32;
+    final double FRONT_CLAW_CLOSED = 0.31;
     final double BACK_CLAW_OPENED = 0.1;
     final double BACK_CLAW_CLOSED = 0.33;
-    final int ARM_POS_UP = -255;
+    final int ARM_POS_UP = -180;
     final int ARM_POS_DOWN = -750;
     final int ARM_POS_TILT = -1310;
+    final int SLIDES_BUCKET_DOWN = 0;
+    final int SLIDES_BUCKET_LOW = 1730;
+    final int SLIDES_BUCKET_HIGH = 3170;
     final int SLIDES_SPECIMEN_DOWN = 0;
-    final int SLIDES_SPECIMEN_TRANSFER = 135;
-    final int SLIDES_SPECIMEN_PREP_HANG = 1000;
-    final int SLIDES_SPECIMEN_HANG = 1600;
-    final int SLIDES_ROBOT_HANG = 1500;
+    final int SLIDES_SPECIMEN_TRANSFER = 590;
+    final int SLIDES_SPECIMEN_PREP_HANG = 1450;
+    final int SLIDES_ROBOT_HANG = 1450;
     final double FRONT_WRIST_HORIZONTAL = 0.61;
     final double STOPPER1_DOWN = 0.7;
     final double STOPPER2_DOWN = 0.74;    // offset seems slightly different on 2
@@ -162,14 +166,18 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
         leftBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
         rightFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
         rightBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+        slideR.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+        slideL.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+        armHinge.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
         //brake motors
         leftFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         leftBackDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         rightFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         rightBackDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        slideR.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        slideL.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        armHinge.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        //slideR.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        //slideL.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        //armHinge.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
         //slide stuff
         slideR.setTargetPosition(0);
         slideL.setTargetPosition(0);
@@ -215,13 +223,13 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
         tongue.setDirection(Servo.Direction.REVERSE);
         wrist.setPosition(0.48);
         claw.setPosition(FRONT_CLAW_OPENED);
-        backWrist.setPosition(0.16);
-        backClaw.setPosition(BACK_CLAW_CLOSED);
+        backWrist.setPosition(0.14);
+        backClaw.setPosition(BACK_CLAW_OPENED);
         rotWrist.setPosition(rotWristPos);
         stopper1.setDirection(Servo.Direction.FORWARD);
-        stopper1.setPosition(STOPPER1_DOWN);
+        stopper1.setPosition(STOPPER1_UP);
         stopper2.setDirection(Servo.Direction.REVERSE);
-        stopper2.setPosition(STOPPER2_DOWN);
+        stopper2.setPosition(STOPPER2_UP);
         tongue.setPosition(0);
         otos.calibrateImu();
         otos.resetTracking();
@@ -279,7 +287,7 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
             }
 
             double botHeading = Math.toRadians(otos.getPosition().h);
-                    //imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS);
+            //imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS);
 
             // Rotate the movement direction counter to the bot's rotation
             double rotX = xscaled * Math.cos(-botHeading) - yscaled * Math.sin(-botHeading);
@@ -350,22 +358,6 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                 //turns off motors when at 0 position
                 slideR.setMotorDisable();
                 slideL.setMotorDisable();
-                /*  comment out the hang slides -- now using B-Button (i.e. bPressed)
-                if(gamepad2.dpad_down && slideLevel == 0){
-                    slideInput = true;
-                    slideTarget = 0;
-                    slideR.setMotorEnable();
-                    slideL.setMotorEnable();
-                    slideR.setVelocity(5000);
-                    slideL.setVelocity(5000);
-                    slideR.setTargetPosition(slideTarget);
-                    slideL.setTargetPosition(slideTarget);
-                    slideLevel = -1;
-                    telemetry.addData("level:","0");
-                    telemetry.addData("power:", "OFF");
-                    telemetry.update();
-                }
-                */
             }
             if (!slideInput){ //slide input true is dpad clicked
                 if (gamepad2.dpad_up) { //when up on dpad is pressed
@@ -376,7 +368,7 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                     if(slideLevel <= 0){ //low basket
                         slideR.setVelocity(5000);
                         slideL.setVelocity(5000);
-                        slideTarget = 1450; //1360;
+                        slideTarget = SLIDES_BUCKET_LOW;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 1;
@@ -384,7 +376,7 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                         telemetry.update();
                     }
                     else if(slideLevel == 1){
-                        slideTarget = 2750;
+                        slideTarget = SLIDES_BUCKET_HIGH;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 2;
@@ -395,7 +387,7 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                 else if(gamepad2.dpad_down){
                     slideInput = true;
                     if(slideLevel == 2){ //low basket
-                        slideTarget = 1450; //1360;
+                        slideTarget = SLIDES_BUCKET_LOW;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 1;
@@ -403,7 +395,7 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                         telemetry.update();
                     }
                     else if(slideLevel == 1){ //down
-                        slideTarget = 0;
+                        slideTarget = SLIDES_BUCKET_DOWN;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 0;
@@ -411,7 +403,7 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                         telemetry.update();
                     }
                     else if(slideLevel <= 0){
-                        slideTarget = 0;
+                        slideTarget = SLIDES_BUCKET_DOWN;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = -1;
@@ -467,7 +459,10 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                         slideLevel = 1;
                         telemetry.addData("level:","1");
                         telemetry.update();
+                        stopper1.setPosition(STOPPER1_DOWN);
+                        stopper2.setPosition(STOPPER2_DOWN);
                     }
+                    /*  Specimen hang doesn't need to go to 3rd level
                     else if(slideLevel == 1){
                         slideTarget = SLIDES_SPECIMEN_HANG;
                         slideR.setTargetPosition(slideTarget);
@@ -476,26 +471,33 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                         telemetry.addData("level:","2");
                         telemetry.update();
                         sleep(500);
-                        backClaw.setPosition(BACK_CLAW_OPENED);
                     }
+                    */
                 }
                 else if(gamepad2.dpad_down){
                     slideInput = true;
-                    if(slideLevel == 2){ //low basket
+                    if(slideLevel == 2){ // go to prep to hang position
                         slideTarget = SLIDES_SPECIMEN_PREP_HANG;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 1;
                         telemetry.addData("level:","1");
                         telemetry.update();
+                        backClaw.setPosition(BACK_CLAW_OPENED);
                     }
-                    else if(slideLevel == 1){ //down
+                    else if(slideLevel == 1){ // put specimen down onto bar
                         slideTarget = SLIDES_SPECIMEN_DOWN;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 0;
+                        stopper1.setPosition(STOPPER1_UP);
+                        stopper2.setPosition(STOPPER2_UP);
                         telemetry.addData("level:","0");
                         telemetry.update();
+                        while (slideR.getCurrentPosition() > 1000) {
+                            sleep(10);
+                        }
+                        backClaw.setPosition(BACK_CLAW_OPENED);
                     }
                     else if(slideLevel <= 0){
                         slideTarget = SLIDES_SPECIMEN_DOWN;
@@ -522,38 +524,45 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
 
     public void arm(){
         armHinge.setTargetPosition(armTarget);
-        int armSpeed = 20;
-        if (armTarget < -750)
+        int armSpeed = 20;  // normal arm speed
+        if (armTarget < -750) {
+            // set a slow speed near the bottom
             armSpeed = 5;
-        if (armTarget>= -890 && gamepad2.right_stick_y > 0){
-            armHinge.setMotorEnable();
-            armHinge.setVelocity(900);
+        }
+
+        if (gamepad2.right_stick_y > 0){
             // Go down
-            armTarget -= (int)(gamepad2.right_stick_y * armSpeed);
-            if (armTarget<-890){
+            if (armTarget>= -890) {
+                // normal moving down -- stop when we get to horizontal
+                armHinge.setMotorEnable();
+                armHinge.setVelocity(900);
+                armTarget -= (int) (gamepad2.right_stick_y * armSpeed);
+            } else if (hangingMode) {
+                // This is during hang -- allowed to go even further!
+                armHinge.setMotorEnable();
+                armHinge.setPower(1.0);
+                armTarget -= (int) (gamepad2.right_stick_y * armSpeed);
+            } else if (armTarget < -890) {
+                // if not hanging mode -- peg at bottom
                 armTarget = -890;
             }
-            armMoving = true;
-            telemetry.addData("arm pos", armHinge.getCurrentPosition());
-            telemetry.addData("tgt pos", armHinge.getTargetPosition());
-            telemetry.update();
-            //upDown.setVelocity(500*upness);
+
         }
         else if (gamepad2.right_stick_y < 0 && armTarget <=0){
+            // Go up
             armHinge.setMotorEnable();
             armHinge.setVelocity(900);
-            // Go up
             armTarget -= (int)(gamepad2.right_stick_y * armSpeed);
             if (armTarget>0){
                 armTarget = 0;
             }
-            armMoving = true;
-            telemetry.addData("arm pos", armHinge.getCurrentPosition());
-            telemetry.addData("tgt pos", armHinge.getTargetPosition());
-            telemetry.update();
         } else if (gamepad2.right_stick_y==0 && armHinge.getCurrentPosition() > -500) {
+            // turn off motor if arm is up, and not commanded to move
             armHinge.setMotorDisable();
         }
+        telemetry.addData("arm pos", armHinge.getCurrentPosition());
+        telemetry.addData("tgt pos", armHinge.getTargetPosition());
+        telemetry.update();
     }
 
     public void tongue(){
@@ -582,84 +591,20 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                 YButtonDown yButtonDown = new YButtonDown();
                 yButtonDown.start();
             }
-
-            /*if (gamepad2.y && depositMode) { //transfer from front claw to back
-                yPressed = true;
-                //motor first
-                int target = SLIDES_SPECIMEN_TRANSFER;
-                backClaw.setPosition(BACK_CLAW_OPENED);
-                backWrist.setPosition(0.62);
-                rotWrist.setPosition(FRONT_WRIST_HORIZONTAL);
-                rotWristPos = FRONT_WRIST_HORIZONTAL;
-                wrist.setPosition(0.04);
-                tongue.setPosition(0);
-                tonguePos = 0;
-                while((slideR.getCurrentPosition()>(target+10) || slideR.getCurrentPosition()<(target-10)) && opModeIsActive()){
-                    slideR.setVelocity(1000);
-                    slideL.setVelocity(1000);
-                    slideR.setTargetPosition(target);
-                    slideL.setTargetPosition(target);
-                    slideLevel=0;
-                    telemetry.addData("SlideR Pos", slideR.getCurrentPosition());
-                    telemetry.addData("SlideR Tgt", slideR.getTargetPosition());
-                    telemetry.update();
-                }
-                armTarget = ARM_POS_UP;
-                int curPos = armHinge.getCurrentPosition();
-                while ((curPos<(ARM_POS_UP-5) || curPos>(ARM_POS_UP+5))&& opModeIsActive()) {
-                    armHinge.setMotorEnable();
-                    armHinge.setVelocity(1500);
-                    armHinge.setTargetPosition(ARM_POS_UP);
-                    armMoving = true;
-                    curPos = armHinge.getCurrentPosition();
-                }
-                backClaw.setPosition(BACK_CLAW_CLOSED);
-                sleep(300);
-                claw.setPosition(FRONT_CLAW_OPENED);
-                sleep(200);
-                backWrist.setPosition(0);
-                depositMode = false;
-
-                // get slide in prep position
-                slideR.setVelocity(5000);
-                slideL.setVelocity(5000);
-                slideTarget = SLIDES_SPECIMEN_PREP_HANG;
-                slideR.setTargetPosition(slideTarget);
-                slideL.setTargetPosition(slideTarget);
-                slideLevel = 1;
-
-            }
-            else if(gamepad2.y){ //grab mode
-                yPressed = true;
-                tongue.setPosition(0.2);
-                tonguePos = 0.2;
-                wrist.setPosition(0.7);
-                claw.setPosition(FRONT_CLAW_OPENED);
-                armTarget = ARM_POS_DOWN;
-                int curPos = armHinge.getCurrentPosition();
-                while (curPos!=ARM_POS_DOWN && opModeIsActive()) {
-                    armHinge.setMotorEnable();
-                    armHinge.setVelocity(800);
-                    armHinge.setTargetPosition(ARM_POS_DOWN);
-                    armMoving = true;
-                    curPos = armHinge.getCurrentPosition();
-                }
-                depositMode = true;
-            }*/
         }
         else if (!gamepad2.y){
             yPressed = false;
         }
     }
 
-    public void claw(){//open-close
-        if(armHinge.getCurrentPosition()<-500){
+    public void claw(){
+        if(!hangingMode){
+            // grabbing position
             wrist.setPosition(0.7);
         }
-        else if(armHinge.getCurrentPosition()>-500){
-            wrist.setPosition(0.48);
-        }
+
         if(!xPressed){
+            //open-close
             if(gamepad2.x && claw.getPosition()>=(FRONT_CLAW_CLOSED-0.1)){
                 xPressed = true;
                 claw.setPosition(FRONT_CLAW_OPENED);
@@ -680,73 +625,84 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
         if(!bPressed){
             if(gamepad2.b){
                 bPressed = true;
+                if (hangingMode) {
+                    hangingMode = false;
+                } else {
+                    hangingMode = true;
 
-                // arm down first
-                armHinge.setMotorEnable();
-                armHinge.setVelocity(900);
-                armHinge.setTargetPosition(ARM_POS_DOWN);
+                    // arm down first
+                    armHinge.setMotorEnable();
+                    armHinge.setVelocity(900);
+                    armTarget = ARM_POS_DOWN;
+                    armHinge.setTargetPosition(armTarget);
 
-                // Start hang: robot needs to be backed up against submersible.
-                // Roll forward few inches to avoid crashing claw into bar
-                leftFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
-                leftBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
-                rightFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
-                rightBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
-                leftFrontDrive.setTargetPosition(50);
-                leftBackDrive.setTargetPosition(50);
-                rightFrontDrive.setTargetPosition(50);
-                rightBackDrive.setTargetPosition(50);
-                leftFrontDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
-                leftBackDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
-                rightFrontDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
-                rightBackDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
-                leftFrontDrive.setVelocity(200);
-                leftBackDrive.setVelocity(200);
-                rightFrontDrive.setVelocity(200);
-                rightBackDrive.setVelocity(200);
+                    // Start hang: robot needs to be backed up against submersible.
+                    // Roll forward few inches to avoid crashing claw into bar
+                    leftFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    leftBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    rightFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    rightBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    leftFrontDrive.setTargetPosition(50);
+                    leftBackDrive.setTargetPosition(50);
+                    rightFrontDrive.setTargetPosition(50);
+                    rightBackDrive.setTargetPosition(50);
+                    leftFrontDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    leftBackDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    rightFrontDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    rightBackDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    leftFrontDrive.setVelocity(200);
+                    leftBackDrive.setVelocity(200);
+                    rightFrontDrive.setVelocity(200);
+                    rightBackDrive.setVelocity(200);
 
-                // Set the arm and wrist in position to push down
-                tongue.setPosition(0.2);
-                wrist.setPosition(0);
+                    // Set the arm and wrist in position to push down
+                    tongue.setPosition(0.2);
+                    wrist.setPosition(0);
 
-                sleep(1000);
+                    sleep(1000);
 
-                // Put slides up
-                slideTarget = SLIDES_ROBOT_HANG;
-                slideR.setTargetPosition(slideTarget);
-                slideL.setTargetPosition(slideTarget);
-                slideR.setVelocity(5000);
-                slideL.setVelocity(5000);
-                slideLevel = 2;
+                    // Put slides up
+                    slideTarget = SLIDES_ROBOT_HANG;
+                    slideR.setTargetPosition(slideTarget);
+                    slideL.setTargetPosition(slideTarget);
+                    slideR.setVelocity(5000);
+                    slideL.setVelocity(5000);
+                    slideLevel = 1;
 
-                sleep(1500);
+                    sleep(1500);
+                    leftFrontDrive.setTargetPosition(0);
+                    leftBackDrive.setTargetPosition(0);
+                    rightFrontDrive.setTargetPosition(0);
+                    rightBackDrive.setTargetPosition(0);
+                    sleep(1500);
 
-                // Tilt the robot -- push arm down
-                while (armHinge.getCurrentPosition() > (ARM_POS_TILT+50) && opModeIsActive()) {
-                    armHinge.setTargetPosition(ARM_POS_TILT);
-                    armMoving = true;
+                    // Set drive motors back to normal
+                    leftFrontDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+                    leftBackDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+                    rightFrontDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+                    rightBackDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+
+                    // Tilt the robot -- push arm down
+                    while (armHinge.getCurrentPosition() > (ARM_POS_TILT + 50) && opModeIsActive()) {
+                        armHinge.setTargetPosition(ARM_POS_TILT);
+                    }
+
+                    // Retract the slides -- hang the robot
+                    slideTarget = SLIDES_SPECIMEN_DOWN;
+                    slideR.setTargetPosition(slideTarget);
+                    slideL.setTargetPosition(slideTarget);
+                    slideLevel = 0;
+
+                    // Pull the arm up
+                    armHinge.setVelocity(900);
+                    armHinge.setTargetPosition(-50);
+
+                    sleep(5000);
+                    armHinge.setMotorDisable();
+
+                    // Wait for end of match
+                    sleep(30000);
                 }
-
-                // Retract the slides -- hang the robot
-                slideTarget = SLIDES_SPECIMEN_DOWN;
-                slideR.setTargetPosition(slideTarget);
-                slideL.setTargetPosition(slideTarget);
-                slideLevel = 0;
-
-                // Pull the arm up
-                armHinge.setTargetPosition(-50);
-
-                // Stop drive motors
-                leftFrontDrive.setMotorDisable();
-                leftBackDrive.setMotorDisable();
-                rightFrontDrive.setMotorDisable();
-                rightBackDrive.setMotorDisable();
-
-                sleep(5000);
-                armHinge.setMotorDisable();
-
-                // Wait for end of match
-                sleep(30000);
             }
         }
         else{
@@ -825,15 +781,15 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
         public void run() {
             try {
                 //motor first
+                backClaw.setPosition(BACK_CLAW_CLOSED);
                 int target = SLIDES_SPECIMEN_TRANSFER;
-                backClaw.setPosition(BACK_CLAW_OPENED);
-                backWrist.setPosition(0.75);
+                backWrist.setPosition(0.77);
                 rotWrist.setPosition(FRONT_WRIST_HORIZONTAL);
                 rotWristPos = FRONT_WRIST_HORIZONTAL;
                 wrist.setPosition(0.04);
                 tongue.setPosition(0);
                 tonguePos = 0;
-                while ((slideR.getCurrentPosition() > (target + 10) || slideR.getCurrentPosition() < (target - 10)) && opModeIsActive()) {
+                while ((slideR.getCurrentPosition() > (target + 5) || slideR.getCurrentPosition() < (target - 5)) && opModeIsActive()) {
                     slideR.setVelocity(1000);
                     slideL.setVelocity(1000);
                     slideR.setTargetPosition(target);
@@ -843,11 +799,18 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                     telemetry.addData("SlideR Tgt", slideR.getTargetPosition());
                     telemetry.update();
                 }
+                backClaw.setPosition(BACK_CLAW_OPENED);
                 armTarget = ARM_POS_UP;
                 int curPos = armHinge.getCurrentPosition();
-                while ((curPos < (ARM_POS_UP - 5) || curPos > (ARM_POS_UP + 5)) && opModeIsActive()) {
+                while ((curPos < (ARM_POS_UP - 1) || curPos > (ARM_POS_UP + 1)) && opModeIsActive()) {
                     armHinge.setMotorEnable();
-                    armHinge.setVelocity(1500);
+                    if(curPos > -300){
+                        armHinge.setVelocity(300);
+                    }
+                    else if(curPos < -400){
+                        armHinge.setVelocity(1500);
+                    }
+                    armHinge.setVelocity(800);
                     armHinge.setTargetPosition(ARM_POS_UP);
                     armMoving = true;
                     curPos = armHinge.getCurrentPosition();
@@ -856,12 +819,16 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                 sleep(300);
                 claw.setPosition(FRONT_CLAW_OPENED);
                 sleep(200);
-                backWrist.setPosition(0.16);
+                backWrist.setPosition(0.14);
                 depositMode = false;
                 // get slide in prep position
                 slideR.setVelocity(5000);
                 slideL.setVelocity(5000);
-                slideTarget = SLIDES_SPECIMEN_PREP_HANG;
+                if (bucketMode) {
+                    slideTarget = SLIDES_BUCKET_LOW;
+                } else {
+                    slideTarget = SLIDES_SPECIMEN_PREP_HANG;
+                }
                 slideR.setTargetPosition(slideTarget);
                 slideL.setTargetPosition(slideTarget);
                 slideLevel = 1;
@@ -882,8 +849,14 @@ public class FC_POST_AUTO_60_SPEED extends LinearOpMode {
                 wrist.setPosition(0.7);
                 armTarget = ARM_POS_DOWN;
                 int curPos = armHinge.getCurrentPosition();
-                while (curPos != ARM_POS_DOWN && opModeIsActive()) {
+                while ((curPos <= ARM_POS_DOWN - 5 || curPos >= ARM_POS_DOWN + 5) && opModeIsActive()) {
                     armHinge.setMotorEnable();
+                    if(curPos>-350){
+                        armHinge.setVelocity(1600);
+                    }
+                    else{
+                        armHinge.setVelocity(250);
+                    }
                     armHinge.setVelocity(800);
                     armHinge.setTargetPosition(ARM_POS_DOWN);
                     armMoving = true;
