+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MainFieldCentric100Speed.java
@@ -30,17 +30,18 @@
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.hardware.rev.RevBlinkinLedDriver;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
+import com.qualcomm.hardware.sparkfun.SparkFunOTOS;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.hardware.IMU;
+import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 
+import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
 /*
  * This file contains an example of a Linear "OpMode".
@@ -72,7 +73,7 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 
 @TeleOp
 //@Disabled
-public class MainRobotCentric60Speed extends LinearOpMode {
+public class MainFieldCentric100Speed extends LinearOpMode {
 
     // Declare OpMode members for each of the 4 motors.
     private ElapsedTime runtime = new ElapsedTime();
@@ -83,12 +84,14 @@ public class MainRobotCentric60Speed extends LinearOpMode {
     private DcMotorEx slideR = null;
     private DcMotorEx slideL = null;
     private DcMotorEx armHinge = null;
-    private CRServo tongue;
+    private Servo tongue;
     private Servo claw;
     private Servo wrist;
     private Servo backWrist;
     private Servo backClaw;
     private Servo rotWrist;
+    private Servo stopper1;
+    private Servo stopper2;
     int slideTarget;
     boolean slideMoving;
     int slideLevel;
@@ -99,12 +102,39 @@ public class MainRobotCentric60Speed extends LinearOpMode {
     boolean yPressed;
     boolean aPressed;
     boolean depositMode;
+    boolean hangingMode = false;
     boolean bPressed;
     boolean bucketMode;
     boolean guidePressed;
     double rotWristPos;
+    double tonguePos;
+    boolean grabbing;
     RevBlinkinLedDriver blinkinLedDriver;
     RevBlinkinLedDriver.BlinkinPattern pattern = RevBlinkinLedDriver.BlinkinPattern.HOT_PINK;
+    SparkFunOTOS otos;
+
+    // Some constant values
+    final double FRONT_CLAW_OPENED = 0.1;
+    final double FRONT_CLAW_CLOSED = 0.31;
+    final double BACK_CLAW_OPENED = 0.1;
+    final double BACK_CLAW_CLOSED = 0.33;
+    final int ARM_POS_UP = -180;
+    final int ARM_POS_DOWN = -750;
+    final int ARM_POS_TILT = -1310;
+    final int SLIDES_BUCKET_DOWN = 0;
+    final int SLIDES_BUCKET_LOW = 1730;
+    final int SLIDES_BUCKET_HIGH = 3170;
+    final int SLIDES_SPECIMEN_DOWN = 0;
+    final int SLIDES_SPECIMEN_TRANSFER = 590;
+    final int SLIDES_SPECIMEN_PREP_HANG = 1450;
+    final int SLIDES_ROBOT_HANG = 1450;
+    final double FRONT_WRIST_HORIZONTAL = 0.61;
+    final double STOPPER1_DOWN = 0.7;
+    final double STOPPER2_DOWN = 0.74;    // offset seems slightly different on 2
+    final double STOPPER1_UP = 0.0;
+    final double STOPPER2_UP = 0.0;
+
+
 
 
     @Override
@@ -119,14 +149,16 @@ public class MainRobotCentric60Speed extends LinearOpMode {
         slideR = hardwareMap.get(DcMotorEx.class, "slideR");
         slideL = hardwareMap.get(DcMotorEx.class, "slideL");
         armHinge = hardwareMap.get(DcMotorEx.class, "armHinge");
-        tongue = hardwareMap.get(CRServo.class, "tongue");
+        tongue = hardwareMap.get(Servo.class, "tongue");
         claw = hardwareMap.get(Servo.class,"claw");
         wrist = hardwareMap.get(Servo.class,"wrist");
         backWrist = hardwareMap.get(Servo.class, "backWrist");
         backClaw = hardwareMap.get(Servo.class, "backClaw");
         rotWrist = hardwareMap.get(Servo.class, "rotWrist");
-        IMU imu = hardwareMap.get(IMU.class, "imu");
+        stopper1 = hardwareMap.get(Servo.class, "stopper1");
+        stopper2 = hardwareMap.get(Servo.class, "stopper2");
         blinkinLedDriver = hardwareMap.get(RevBlinkinLedDriver.class, "blinkin");
+        otos = hardwareMap.get(SparkFunOTOS.class, "otos");
 
         //reset encoder
         leftFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
@@ -151,10 +183,6 @@ public class MainRobotCentric60Speed extends LinearOpMode {
         slideR.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         slideL.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         armHinge.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
-                RevHubOrientationOnRobot.LogoFacingDirection.UP,
-                RevHubOrientationOnRobot.UsbFacingDirection.FORWARD));
-        imu.initialize(parameters);
         slideTarget = 0;
         slideMoving = false;
         slideLevel = 0;
@@ -168,7 +196,9 @@ public class MainRobotCentric60Speed extends LinearOpMode {
         bPressed = false;
         bucketMode = false;
         guidePressed = false;
-        rotWristPos = 0.64;
+        grabbing = false;
+        rotWristPos = FRONT_WRIST_HORIZONTAL;
+        tonguePos = 0;
 
         // ########################################################################################
         // !!!            IMPORTANT Drive Information. Test your motor directions.            !!!!!
@@ -188,27 +218,46 @@ public class MainRobotCentric60Speed extends LinearOpMode {
         armHinge.setDirection(DcMotorSimple.Direction.REVERSE);
         wrist.setDirection(Servo.Direction.REVERSE);
         backWrist.setDirection(Servo.Direction.REVERSE);
-        wrist.setPosition(0);
-        claw.setPosition(0);
-        backWrist.setPosition(0);
-        backClaw.setPosition(0.35);
-        rotWrist.setPosition(0.64);
+        tongue.setDirection(Servo.Direction.REVERSE);
+        wrist.setPosition(0.48);
+        claw.setPosition(FRONT_CLAW_OPENED);
+        backWrist.setPosition(0.14);
+        backClaw.setPosition(BACK_CLAW_OPENED);
+        rotWrist.setPosition(rotWristPos);
+        stopper1.setDirection(Servo.Direction.FORWARD);
+        stopper1.setPosition(STOPPER1_UP);
+        stopper2.setDirection(Servo.Direction.REVERSE);
+        stopper2.setPosition(STOPPER2_UP);
+        tongue.setPosition(0);
+        otos.calibrateImu();
+        otos.resetTracking();
+        SparkFunOTOS.Pose2D currentPosition = new SparkFunOTOS.Pose2D(0, 0, 0);
+        otos.setPosition(currentPosition);
+
+        // Retrieve the IMU from the hardware map
+        IMU imu = hardwareMap.get(IMU.class, "imu");
+        // Adjust the orientation parameters to match your robot
+        IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
+                RevHubOrientationOnRobot.LogoFacingDirection.UP,
+                RevHubOrientationOnRobot.UsbFacingDirection.FORWARD));
+        // Without this, the REV Hub's orientation is assumed to be logo up / USB forward
+        imu.initialize(parameters);
 
 
-                // Wait for the game to start (driver presses START)
+
+        // Wait for the game to start (driver presses START)
         telemetry.addData("Status", "Initialized");
         telemetry.update();
 
         waitForStart();
         runtime.reset();
 
-        pattern = RevBlinkinLedDriver.BlinkinPattern.RED;
-        blinkinLedDriver.setPattern(pattern);
-
         leftFrontDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
         leftBackDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
         rightFrontDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
         rightBackDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+        imu.resetYaw();
+
         // run until the end of the match (driver presses STOP)
         while (opModeIsActive()) {
             /*scaling equation:
@@ -219,40 +268,58 @@ public class MainRobotCentric60Speed extends LinearOpMode {
              * k: max power constant
              * n: reduces sensitivity for smaller values of x - greater value of n makes smaller values less powerful
              * */
-            double y = -(gamepad1.left_stick_y ) * .6; // Remember, Y stick value is reversed
+            double y = -gamepad1.left_stick_y; // Remember, Y stick value is reversed
             double yscaled = y != 0 ? Math.signum(y) * Math.pow(Math.abs(y), 2) : 0;
-            double x = gamepad1.left_stick_x * 1.1 * .6; // Counteract imperfect strafing
+            double x = gamepad1.left_stick_x;
             double xscaled = x != 0 ? Math.signum(x) * Math.pow(Math.abs(x), 2) : 0;
-            double rx = gamepad1.right_stick_x * .6;
+            double rx = gamepad1.right_stick_x;
             double rxscaled = rx != 0 ? Math.signum(rx) * Math.pow(Math.abs(rx), 2) : 0;
-            double denominator = Math.max(Math.abs(y) + Math.abs(x) + Math.abs(rx), 1);
-            double rightBackPower = (yscaled + xscaled - rxscaled) / denominator;
-            double leftBackPower = (yscaled - xscaled + rxscaled) / denominator;
-            double leftFrontPower = (yscaled + xscaled + rxscaled) / denominator;
-            double rightFrontPower = (yscaled - xscaled - rxscaled) / denominator;
 
-            //Drivetrain (Gamepad1 left + right joystick)
-            leftFrontDrive.setPower(leftFrontPower);
-            leftBackDrive.setPower(leftBackPower);
-            rightFrontDrive.setPower(rightFrontPower);
-            rightBackDrive.setPower(rightBackPower);
+            // This button choice was made so that it is hard to hit on accident,
+            // it can be freely changed based on preference.
+            // The equivalent button is start on Xbox-style controllers.
+            if (gamepad1.x) {
+                otos.calibrateImu();
+                otos.resetTracking();
+                otos.setPosition(currentPosition);
+            }
 
-            // Show the elapsed game time and wheel power.
+            double botHeading = Math.toRadians(otos.getPosition().h);
+            //imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS);
 
-            /*telemetry.addData("Status", "Run Time: " + runtime.toString());
-            telemetry.addData("Angle", imu.getRobotYawPitchRollAngles());
-            telemetry.addData("Front left/Right", "%4.2f, %4.2f", leftFrontPower, rightFrontPower);
-            telemetry.addData("Back  left/Right", "%4.2f, %4.2f", leftBackPower, rightBackPower);
-            telemetry.update();*/
-            slide();
-            arm();
-            tongue();
-            grabDeposit();
-            claw();
-            lights();
-            tiltHang();
-            backClaw();
-            rotWrist();
+            // Rotate the movement direction counter to the bot's rotation
+            double rotX = xscaled * Math.cos(-botHeading) - yscaled * Math.sin(-botHeading);
+            double rotY = xscaled * Math.sin(-botHeading) + yscaled * Math.cos(-botHeading);
+
+            //rotX = rotX * 1.1;  // Counteract imperfect strafing
+
+            // Denominator is the largest motor power (absolute value) or 1
+            // This ensures all the powers maintain the same ratio,
+            // but only if at least one is out of the range [-1, 1]
+            double denominator = Math.max(Math.abs(rotY) + Math.abs(rotX) + Math.abs(rx), 1);
+            double frontLeftPower = (rotY + rotX + rxscaled) / denominator;
+            double backLeftPower = (rotY - rotX + rxscaled) / denominator;
+            double frontRightPower = (rotY - rotX - rxscaled) / denominator;
+            double backRightPower = (rotY + rotX - rxscaled) / denominator;
+
+            leftFrontDrive.setPower(frontLeftPower);
+            leftBackDrive.setPower(backLeftPower);
+            rightFrontDrive.setPower(frontRightPower);
+            rightBackDrive.setPower(backRightPower);
+
+            if(!grabbing){
+                slide();
+                arm();
+                tongue();
+                grabDeposit();
+                claw();
+                lights();
+                tiltHang();
+                backClaw();
+                rotWrist();
+                stoppers();
+            }
+
             telemetry.addData("slide height", slideR.getCurrentPosition());
             if(bucketMode){
                 telemetry.addData("Mode: ", "Bucket/Hang");
@@ -260,6 +327,8 @@ public class MainRobotCentric60Speed extends LinearOpMode {
             else{
                 telemetry.addData("Mode: ", "Specimen");
             }
+            telemetry.addData("otos heading:", Math.toRadians(otos.getPosition().h));
+            telemetry.addData("imu output: ",imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS));
             telemetry.addData("arm pos", armHinge.getCurrentPosition());
             telemetry.addData("tgt pos", armHinge.getTargetPosition());
             telemetry.update();
@@ -287,20 +356,6 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                 //turns off motors when at 0 position
                 slideR.setMotorDisable();
                 slideL.setMotorDisable();
-                if(gamepad2.dpad_down && slideLevel == 0){
-                    slideInput = true;
-                    slideTarget = 0;
-                    slideR.setMotorEnable();
-                    slideL.setMotorEnable();
-                    slideR.setVelocity(5000);
-                    slideL.setVelocity(5000);
-                    slideR.setTargetPosition(slideTarget);
-                    slideL.setTargetPosition(slideTarget);
-                    slideLevel = -1;
-                    telemetry.addData("level:","0");
-                    telemetry.addData("power:", "OFF");
-                    telemetry.update();
-                }
             }
             if (!slideInput){ //slide input true is dpad clicked
                 if (gamepad2.dpad_up) { //when up on dpad is pressed
@@ -311,15 +366,15 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                     if(slideLevel <= 0){ //low basket
                         slideR.setVelocity(5000);
                         slideL.setVelocity(5000);
-                        slideTarget = 1450; //1360;
+                        slideTarget = SLIDES_BUCKET_LOW;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 1;
                         telemetry.addData("level:","1");
                         telemetry.update();
                     }
-                    else if(slideLevel == 1 && armHinge.getCurrentPosition()> -300){ //high basket
-                        slideTarget = 2750;
+                    else if(slideLevel == 1){
+                        slideTarget = SLIDES_BUCKET_HIGH;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 2;
@@ -330,7 +385,7 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                 else if(gamepad2.dpad_down){
                     slideInput = true;
                     if(slideLevel == 2){ //low basket
-                        slideTarget = 1450; //1360;
+                        slideTarget = SLIDES_BUCKET_LOW;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 1;
@@ -338,7 +393,7 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                         telemetry.update();
                     }
                     else if(slideLevel == 1){ //down
-                        slideTarget = 0;
+                        slideTarget = SLIDES_BUCKET_DOWN;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 0;
@@ -346,7 +401,7 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                         telemetry.update();
                     }
                     else if(slideLevel <= 0){
-                        slideTarget = 0;
+                        slideTarget = SLIDES_BUCKET_DOWN;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = -1;
@@ -356,7 +411,7 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                 }
             }
             else {
-                if (!gamepad2.dpad_down && !gamepad2.dpad_up && !yPressed){ //when arrow button (dpad up/down) is released
+                if (!gamepad2.dpad_down && !gamepad2.dpad_up){ //when arrow button (dpad up/down) is released
                     slideInput = false;
                     slideMoving = false;
                 /*telemetry.addData("Right position:", slideR.getCurrentPosition());
@@ -367,13 +422,14 @@ public class MainRobotCentric60Speed extends LinearOpMode {
             }
         }
         else{
+            // Specimen mode
             if(slideR.getCurrentPosition() < 20 && slideLevel == 0){
                 //turns off motors when at 0 position
                 slideR.setMotorDisable();
                 slideL.setMotorDisable();
                 if(gamepad2.dpad_down && slideLevel == 0){
                     slideInput = true;
-                    slideTarget = 0;
+                    slideTarget = SLIDES_SPECIMEN_DOWN;
                     slideR.setMotorEnable();
                     slideL.setMotorEnable();
                     slideR.setVelocity(5000);
@@ -395,42 +451,54 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                     if(slideLevel <= 0){ //low basket
                         slideR.setVelocity(5000);
                         slideL.setVelocity(5000);
-                        slideTarget = 850;
+                        slideTarget = SLIDES_SPECIMEN_PREP_HANG;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 1;
                         telemetry.addData("level:","1");
                         telemetry.update();
+                        stopper1.setPosition(STOPPER1_DOWN);
+                        stopper2.setPosition(STOPPER2_DOWN);
                     }
-                    else if(slideLevel == 1 && armHinge.getCurrentPosition()> -300){ //high basket
-                        slideTarget = 1350;
+                    /*  Specimen hang doesn't need to go to 3rd level
+                    else if(slideLevel == 1){
+                        slideTarget = SLIDES_SPECIMEN_HANG;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 2;
                         telemetry.addData("level:","2");
                         telemetry.update();
+                        sleep(500);
                     }
+                    */
                 }
                 else if(gamepad2.dpad_down){
                     slideInput = true;
-                    if(slideLevel == 2){ //low basket
-                        slideTarget = 850;
+                    if(slideLevel == 2){ // go to prep to hang position
+                        slideTarget = SLIDES_SPECIMEN_PREP_HANG;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 1;
                         telemetry.addData("level:","1");
                         telemetry.update();
+                        backClaw.setPosition(BACK_CLAW_OPENED);
                     }
-                    else if(slideLevel == 1){ //down
-                        slideTarget = 0;
+                    else if(slideLevel == 1){ // put specimen down onto bar
+                        slideTarget = SLIDES_SPECIMEN_DOWN;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = 0;
+                        stopper1.setPosition(STOPPER1_UP);
+                        stopper2.setPosition(STOPPER2_UP);
                         telemetry.addData("level:","0");
                         telemetry.update();
+                        while (slideR.getCurrentPosition() > 1000) {
+                            sleep(10);
+                        }
+                        backClaw.setPosition(BACK_CLAW_OPENED);
                     }
                     else if(slideLevel <= 0){
-                        slideTarget = 0;
+                        slideTarget = SLIDES_SPECIMEN_DOWN;
                         slideR.setTargetPosition(slideTarget);
                         slideL.setTargetPosition(slideTarget);
                         slideLevel = -1;
@@ -440,7 +508,7 @@ public class MainRobotCentric60Speed extends LinearOpMode {
                 }
             }
             else {
-                if (!gamepad2.dpad_down && !gamepad2.dpad_up && !yPressed){ //when arrow button (dpad up/down) is released
+                if (!gamepad2.dpad_down && !gamepad2.dpad_up){ //when arrow button (dpad up/down) is released
                     slideInput = false;
                     slideMoving = false;
                 /*telemetry.addData("Right position:", slideR.getCurrentPosition());
@@ -454,134 +522,94 @@ public class MainRobotCentric60Speed extends LinearOpMode {
 
     public void arm(){
         armHinge.setTargetPosition(armTarget);
-        //armHinge.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        if (armTarget>= -890 && gamepad2.right_stick_y > 0 && slideLevel<2){
-            armHinge.setVelocity(900);
+        int armSpeed = 20;  // normal arm speed
+        if (armTarget < -750) {
+            // set a slow speed near the bottom
+            armSpeed = 5;
+        }
+
+        if (gamepad2.right_stick_y > 0){
             // Go down
-            armTarget -= (int)(gamepad2.right_stick_y*20);
-            if (armTarget<-890){
+            if (armTarget>= -890) {
+                // normal moving down -- stop when we get to horizontal
+                armHinge.setMotorEnable();
+                armHinge.setVelocity(900);
+                armTarget -= (int) (gamepad2.right_stick_y * armSpeed);
+            } else if (hangingMode) {
+                // This is during hang -- allowed to go even further!
+                armHinge.setMotorEnable();
+                armHinge.setPower(1.0);
+                armTarget -= (int) (gamepad2.right_stick_y * armSpeed);
+            } else if (armTarget < -890) {
+                // if not hanging mode -- peg at bottom
                 armTarget = -890;
             }
-            armMoving = true;
-            telemetry.addData("arm pos", armHinge.getCurrentPosition());
-            telemetry.addData("tgt pos", armHinge.getTargetPosition());
-            telemetry.update();
-            //upDown.setVelocity(500*upness);
+
         }
-        else if (gamepad2.right_stick_y < 0 && armTarget <=0 && slideLevel<2){
-            armHinge.setVelocity(900);
+        else if (gamepad2.right_stick_y < 0 && armTarget <=0){
             // Go up
-            armTarget -= (int)(gamepad2.right_stick_y*20);
+            armHinge.setMotorEnable();
+            armHinge.setVelocity(900);
+            armTarget -= (int)(gamepad2.right_stick_y * armSpeed);
             if (armTarget>0){
                 armTarget = 0;
             }
-            armMoving = true;
-            telemetry.addData("arm pos", armHinge.getCurrentPosition());
-            telemetry.addData("tgt pos", armHinge.getTargetPosition());
-            telemetry.update();
-            //upDown.setVelocity(500*upness);
+        } else if (gamepad2.right_stick_y==0 && armHinge.getCurrentPosition() > -500) {
+            // turn off motor if arm is up, and not commanded to move
+            armHinge.setMotorDisable();
         }
-        else {
-            if (armMoving){
-                armTarget = armHinge.getCurrentPosition();
-                armMoving = false;
-                // STOP
-                //upDown.setVelocity(0);
-                //Do nothing for now - holding position
-            }
-        }
+        telemetry.addData("arm pos", armHinge.getCurrentPosition());
+        telemetry.addData("tgt pos", armHinge.getTargetPosition());
+        telemetry.update();
     }
 
     public void tongue(){
-        if (gamepad2.left_bumper) {
-            tongue.setPower(1);
-        } else if (gamepad2.right_bumper) {
-            tongue.setPower(-1);
-        } else {
-            tongue.setPower(0);
+        if (gamepad2.left_bumper && tonguePos > 0) {
+            tonguePos -= 0.01;
+        } else if (gamepad2.right_bumper && tonguePos <0.37) {
+            tonguePos += 0.01;
         }
+        tongue.setPosition(tonguePos);
+
     }
 
     public void grabDeposit(){
-        if(!yPressed) {
-            if (gamepad2.y && depositMode) { //transfer from front claw to back
+        if(!yPressed &&!grabbing) {
+
+            if (gamepad2.y &&depositMode){
                 yPressed = true;
-                //motor first
-                telemetry.addData("Sdlie r", slideR.getCurrentPosition());
-                telemetry.addData("tgt", slideR.getTargetPosition());
-                telemetry.update();
-                while((slideR.getCurrentPosition()>190 || slideR.getCurrentPosition()<170) && opModeIsActive()){
-                    slideR.setVelocity(1000);
-                    slideL.setVelocity(1000);
-                    slideR.setTargetPosition(180);
-                    slideL.setTargetPosition(180);
-                    slideLevel=0;
-                }
-                backClaw.setPosition(0.35);
-                backWrist.setPosition(0.65);
-                rotWrist.setPosition(0.64);
-                rotWristPos = 0.64;
-                while (armHinge.getCurrentPosition() < 0 && opModeIsActive()) {
-                    armHinge.setVelocity(900);
-                    armHinge.setTargetPosition(0);
-                    tongue.setPower(1);
-                    armMoving = true;
-                }
-                tongue.setPower(0);
-                while(wrist.getPosition()!= 0 && opModeIsActive()){
-                    wrist.setPosition(0);
-                }
-                sleep(600);
-                backClaw.setPosition(0);
-                sleep(300);
-                claw.setPosition(0);
-                sleep(300);
-                backWrist.setPosition(0);
-                depositMode = false;
+                grabbing = true;
+                YButtonUp yButtonUp = new YButtonUp();
+                yButtonUp.start();
             }
-            else if(gamepad2.y){ //grab mode
+            else if(gamepad2.y){
                 yPressed = true;
-                /*while(slideR.getCurrentPosition()>1360){
-                    slideR.setVelocity(1000);
-                    slideL.setVelocity(1000);
-                    slideR.setTargetPosition(210);
-                    slideL.setTargetPosition(210);
-                    slideLevel=0;
-                }*/
-                while (armHinge.getCurrentPosition() > -550 && opModeIsActive()) {
-                    armHinge.setVelocity(900);
-                    armHinge.setTargetPosition(-550);
-                    if(armHinge.getCurrentPosition() < -450){
-                        tongue.setPower(-1);
-                    }
-                    else{
-                        tongue.setPower(0);
-                    }
-                    armMoving = true;
-                }
-                wrist.setPosition(0.62);
-                claw.setPosition(0);
-                depositMode = true;
+                grabbing = true;
+                wrist.setPosition(0.7);
+                YButtonDown yButtonDown = new YButtonDown();
+                yButtonDown.start();
             }
         }
         else if (!gamepad2.y){
             yPressed = false;
-            armMoving = false;
         }
     }
 
-    public void claw(){//open-close
-        if(armHinge.getCurrentPosition()<-300 && !bPressed){
-            wrist.setPosition(0.62);
+    public void claw(){
+        if(!hangingMode){
+            // grabbing position
+            wrist.setPosition(0.7);
         }
+
         if(!xPressed){
-            if(gamepad2.x && claw.getPosition()==0.34){
+            //open-close
+            if(gamepad2.x && claw.getPosition()>=(FRONT_CLAW_CLOSED-0.1)){
                 xPressed = true;
-                claw.setPosition(0);
+                claw.setPosition(FRONT_CLAW_OPENED);
             }
             else if(gamepad2.x){
                 xPressed = true;
-                claw.setPosition(0.34);
+                claw.setPosition(FRONT_CLAW_CLOSED);
             }
         }
         else{
@@ -595,26 +623,84 @@ public class MainRobotCentric60Speed extends LinearOpMode {
         if(!bPressed){
             if(gamepad2.b){
                 bPressed = true;
-                wrist.setPosition(0);
-                while (armHinge.getCurrentPosition() > -1400 && opModeIsActive()) {
-                    armHinge.setTargetPosition(-1400);
-                    armMoving = true;
+                if (hangingMode) {
+                    hangingMode = false;
+                } else {
+                    hangingMode = true;
+
+                    // arm down first
+                    armHinge.setMotorEnable();
+                    armHinge.setVelocity(900);
+                    armTarget = ARM_POS_DOWN;
+                    armHinge.setTargetPosition(armTarget);
+
+                    // Start hang: robot needs to be backed up against submersible.
+                    // Roll forward few inches to avoid crashing claw into bar
+                    leftFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    leftBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    rightFrontDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    rightBackDrive.setMode(DcMotorEx.RunMode.STOP_AND_RESET_ENCODER);
+                    leftFrontDrive.setTargetPosition(50);
+                    leftBackDrive.setTargetPosition(50);
+                    rightFrontDrive.setTargetPosition(50);
+                    rightBackDrive.setTargetPosition(50);
+                    leftFrontDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    leftBackDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    rightFrontDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    rightBackDrive.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
+                    leftFrontDrive.setVelocity(200);
+                    leftBackDrive.setVelocity(200);
+                    rightFrontDrive.setVelocity(200);
+                    rightBackDrive.setVelocity(200);
+
+                    // Set the arm and wrist in position to push down
+                    tongue.setPosition(0.2);
+                    wrist.setPosition(0);
+
+                    sleep(1000);
+
+                    // Put slides up
+                    slideTarget = SLIDES_ROBOT_HANG;
+                    slideR.setTargetPosition(slideTarget);
+                    slideL.setTargetPosition(slideTarget);
+                    slideR.setVelocity(5000);
+                    slideL.setVelocity(5000);
+                    slideLevel = 1;
+
+                    sleep(1500);
+                    leftFrontDrive.setTargetPosition(0);
+                    leftBackDrive.setTargetPosition(0);
+                    rightFrontDrive.setTargetPosition(0);
+                    rightBackDrive.setTargetPosition(0);
+                    sleep(1500);
+
+                    // Set drive motors back to normal
+                    leftFrontDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+                    leftBackDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+                    rightFrontDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+                    rightBackDrive.setMode(DcMotorEx.RunMode.RUN_USING_ENCODER);
+
+                    // Tilt the robot -- push arm down
+                    while (armHinge.getCurrentPosition() > (ARM_POS_TILT + 50) && opModeIsActive()) {
+                        armHinge.setTargetPosition(ARM_POS_TILT);
+                    }
+
+                    // Retract the slides -- hang the robot
+                    slideTarget = SLIDES_SPECIMEN_DOWN;
+                    slideR.setTargetPosition(slideTarget);
+                    slideL.setTargetPosition(slideTarget);
+                    slideLevel = 0;
+
+                    // Pull the arm up
+                    armHinge.setVelocity(900);
+                    armHinge.setTargetPosition(-50);
+
+                    sleep(5000);
+                    armHinge.setMotorDisable();
+
+                    // Wait for end of match
+                    sleep(30000);
                 }
-                int temp = (int)(getRuntime());
-                while((getRuntime() < temp+1) && opModeIsActive()){
-                    tongue.setPower(-1);
-                }
-                temp = (int)(getRuntime());
-                while((getRuntime()< temp+1) && opModeIsActive()){
-                    leftFrontDrive.setPower(-1);
-                    leftBackDrive.setPower(-1);
-                    rightFrontDrive.setPower(-1);
-                    rightBackDrive.setPower(-1);
-                }
-                leftFrontDrive.setPower(0);
-                leftBackDrive.setPower(0);
-                rightFrontDrive.setPower(0);
-                rightBackDrive.setPower(0);
             }
         }
         else{
@@ -626,13 +712,13 @@ public class MainRobotCentric60Speed extends LinearOpMode {
 
     public void backClaw(){
         if(!aPressed){
-            if(gamepad2.a && backClaw.getPosition()==0){
+            if(gamepad2.a && backClaw.getPosition()>=(BACK_CLAW_CLOSED-0.1)){
                 aPressed = true;
-                backClaw.setPosition(0.35);
+                backClaw.setPosition(BACK_CLAW_OPENED);
             }
             else if(gamepad2.a){
                 aPressed = true;
-                backClaw.setPosition(0);
+                backClaw.setPosition(BACK_CLAW_CLOSED);
             }
         }
         else{
@@ -642,14 +728,25 @@ public class MainRobotCentric60Speed extends LinearOpMode {
         }
     }
 
+    public void stoppers(){
+        if(gamepad2.left_stick_y==-1.0) {
+            stopper1.setPosition(STOPPER1_UP);
+            stopper2.setPosition(STOPPER2_UP);
+        }
+        else if(gamepad2.left_stick_y==1.0) {
+            stopper1.setPosition(STOPPER1_DOWN);
+            stopper2.setPosition(STOPPER2_DOWN);
+        }
+    }
+
     public void rotWrist(){
-        if (rotWrist.getPosition() < 1 && gamepad2.right_stick_x > 0){
+        if (rotWrist.getPosition() < 1 && gamepad2.left_stick_x < 0){
             // Go right
             rotWristPos += 0.01;
             rotWrist.setPosition(rotWristPos);
             //upDown.setVelocity(500*upness);
         }
-        else if (gamepad2.right_stick_x < 0 && rotWrist.getPosition() > 0){
+        else if (gamepad2.left_stick_x > 0 && rotWrist.getPosition() > 0){
             // Go left
             rotWristPos -= 0.01;
             rotWrist.setPosition(rotWristPos);
@@ -662,10 +759,112 @@ public class MainRobotCentric60Speed extends LinearOpMode {
             pattern = RevBlinkinLedDriver.BlinkinPattern.CP1_STROBE;
             blinkinLedDriver.setPattern(pattern);
         }
-        if(runtime.seconds()<=105 && runtime.seconds()>=104){
+        else if(runtime.seconds()<=105 && runtime.seconds()>=104){
             pattern = RevBlinkinLedDriver.BlinkinPattern.SHOT_WHITE;
             blinkinLedDriver.setPattern(pattern);
         }
+        else if(bucketMode){
+            pattern = RevBlinkinLedDriver.BlinkinPattern.HOT_PINK;
+            blinkinLedDriver.setPattern(pattern);
+        }
+        else{
+            pattern = RevBlinkinLedDriver.BlinkinPattern.BLUE;
+            blinkinLedDriver.setPattern(pattern);
+        }
 
+    }
+
+    public class YButtonUp extends Thread{
+        @Override
+        public void run() {
+            try {
+                //motor first
+                backClaw.setPosition(BACK_CLAW_CLOSED);
+                int target = SLIDES_SPECIMEN_TRANSFER;
+                backWrist.setPosition(0.77);
+                rotWrist.setPosition(FRONT_WRIST_HORIZONTAL);
+                rotWristPos = FRONT_WRIST_HORIZONTAL;
+                wrist.setPosition(0.04);
+                tongue.setPosition(0);
+                tonguePos = 0;
+                while ((slideR.getCurrentPosition() > (target + 5) || slideR.getCurrentPosition() < (target - 5)) && opModeIsActive()) {
+                    slideR.setVelocity(1000);
+                    slideL.setVelocity(1000);
+                    slideR.setTargetPosition(target);
+                    slideL.setTargetPosition(target);
+                    slideLevel = 0;
+                    telemetry.addData("SlideR Pos", slideR.getCurrentPosition());
+                    telemetry.addData("SlideR Tgt", slideR.getTargetPosition());
+                    telemetry.update();
+                }
+                backClaw.setPosition(BACK_CLAW_OPENED);
+                armTarget = ARM_POS_UP;
+                int curPos = armHinge.getCurrentPosition();
+                while ((curPos < (ARM_POS_UP - 1) || curPos > (ARM_POS_UP + 1)) && opModeIsActive()) {
+                    armHinge.setMotorEnable();
+                    if(curPos > -300){
+                        armHinge.setVelocity(300);
+                    }
+                    else if(curPos < -400){
+                        armHinge.setVelocity(1500);
+                    }
+                    armHinge.setVelocity(800);
+                    armHinge.setTargetPosition(ARM_POS_UP);
+                    armMoving = true;
+                    curPos = armHinge.getCurrentPosition();
+                }
+                backClaw.setPosition(BACK_CLAW_CLOSED);
+                sleep(300);
+                claw.setPosition(FRONT_CLAW_OPENED);
+                sleep(200);
+                backWrist.setPosition(0.14);
+                depositMode = false;
+                // get slide in prep position
+                slideR.setVelocity(5000);
+                slideL.setVelocity(5000);
+                if (bucketMode) {
+                    slideTarget = SLIDES_BUCKET_LOW;
+                } else {
+                    slideTarget = SLIDES_SPECIMEN_PREP_HANG;
+                }
+                slideR.setTargetPosition(slideTarget);
+                slideL.setTargetPosition(slideTarget);
+                slideLevel = 1;
+                grabbing = false;
+            } catch (Exception ex) {
+
+            }
+        }
+    }
+
+    public class YButtonDown extends Thread{
+        @Override
+        public void run() {
+            try {
+                tongue.setPosition(0.2);
+                tonguePos = 0.2;
+                claw.setPosition(FRONT_CLAW_OPENED);
+                wrist.setPosition(0.7);
+                armTarget = ARM_POS_DOWN;
+                int curPos = armHinge.getCurrentPosition();
+                while ((curPos <= ARM_POS_DOWN - 5 || curPos >= ARM_POS_DOWN + 5) && opModeIsActive()) {
+                    armHinge.setMotorEnable();
+                    if(curPos>-350){
+                        armHinge.setVelocity(1600);
+                    }
+                    else{
+                        armHinge.setVelocity(250);
+                    }
+                    armHinge.setVelocity(800);
+                    armHinge.setTargetPosition(ARM_POS_DOWN);
+                    armMoving = true;
+                    curPos = armHinge.getCurrentPosition();
+                }
+                depositMode = true;
+                grabbing = false;
+            } catch (Exception ex) {
+
+            }
+        }
     }
 }
